refactor: split clang attribute normalization
